Move extendedquantifiers package out of preprocessors, add InstrumentationLoop class.
